YSP-539: Bypass field_metatags if they don't exist (#8)
* fix(YSP-539): bypass metatags if they don't exist

This is a first pass, where if the metatags field is not defined on the
nodes, we do not use it.  Otherwise, we will use it.

Next up will be defining the metatags name to check against, if any.

* fix(YSP-539): provide next page if no page exists

When visiting the first callback URL, page did not exist, so it would
show the next url as being page=1.  This now allows it to show 2 so that
we don't re-index the same data.

* fix(YSP-539): bypass nodes that can't render

In our tests we ran into a node that wouldn't successfully render.  This
bypasses this case and returns no data.