fix(YSP-592): set request title to node title
When meta blocks render, they attempt to use the request object to get
the title of the route.  In the case of visiting a page, this works as
intended, but in the case of our AI content endpoint, it resolves to
that endpoint, resulting in the title always being 'Content Feed', since
that is the title of that endpoint.

This temporarily sets the route's title to the node's title so that the
render can successfully occur using the correct title.