feat(YSP-672): new media does not ai index by default
While we allow new pages to be indexed by default, it did not make sense
to always opt-out on media that is uploaded.  This adds the feature that
when they're inserting a new media that has ai metatags enabled, it will
default the media item to be disabled by default.