fix(YSP-672): use media URL for documentUrl
Instead of placaing the URL inside documentContent, which doesn't quite
make sense, place it in the documentUrl.  This allows citations to link
directly to the PDF, and the edit page it was referencing before really
is useless to the users since they won't have access to visit that link
anwyay.